pyproject.toml
[project]
name = "products"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0,<7.0",
    "pillow>=10.0.0",
    "python-decouple>=3.8",
    "gunicorn>=21.0.0",
    "whitenoise>=6.6.0",
    "psycopg2-binary>=2.9.9",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "django-allauth>=0.57.0",
    "pyjwt>=2.8.0",
    "django-debug-toolbar>=4.2.0"
]

# [tool.poetry]
# packages = [{include = "products", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
